main-appfiles/backend/requirements.txt
Flask==2.3.3
Flask-Cors==4.0.0
Flask-SQLAlchemy==3.1.1
google-generativeai==0.3.1
imaplib2==3.6
scikit-learn==1.3.2
pandas==2.1.1
numpy==1.26.0
python-dotenv==1.0.0
requests==2.31.0
email-validator==2.1.0.post1
Flask-Login==0.6.2
authlib==1.2.1
itsdangerous==2.1.2
joblib==1.3.2
gunicorn==21.2.0
psycopg2-binary==2.9.9
werkzeug==2.3.7
beautifulsoup4==4.12.2 